+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   `react.Procedure` as the type of the callback in the top-level rendering methods.
 * **\[dom\]** Introduce a functional interface `react.dom.ReactDOM.BatchedUpdatesFn` to replace the use of
   `react.Procedure` as the type of the callback for the `ReactDOM.batchedUpdates(BatchedUpdatesFn)` methods.
+* **\[dom\]** Introduce several more utility methods to the dom factory class that accepts primitive values
+  as content.
 
 ### [v0.99](https://github.com/react4j/react4j/tree/v0.99) (2018-10-05)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.98...v0.99)